+++ src/Magento/MagentoCloud/Process/Deploy/PreDeploy.php
@@ -9,8 +9,8 @@ use Magento\MagentoCloud\Process\ProcessInterface;
 use Magento\MagentoCloud\Shell\ShellInterface;
 use Magento\MagentoCloud\Filesystem\Driver\File;
 use Magento\MagentoCloud\Config\Environment;
-use Magento\MagentoCloud\Utils\ComponentInfo;
-use Magento\MagentoCloud\Utils\StaticContentCleaner;
+use Magento\MagentoCloud\Util\ComponentInfo;
+use Magento\MagentoCloud\Util\StaticContentCleaner;
 use Psr\Log\LoggerInterface;
 
 /**
